refactor: 서비스 쿼리 유틸 클래스로 분리, DTO 변환 유틸 클래스로 분리, 커스텀 예외처리
