+++ concert-service/src/main/java/com/eatpizzaquickly/concertservice/exception/detail/AutocompleteException.java
@@ -1,0 +1,7 @@
+package com.eatpizzaquickly.concertservice.exception.detail;
+
+public class AutocompleteException extends RuntimeException {
+    public AutocompleteException(String message) {
+        super(message);
+    }
+}
