+++ concert-service/src/main/java/com/eatpizzaquickly/concertservice/exception/detail/ConcertSearchException.java
@@ -1,0 +1,7 @@
+package com.eatpizzaquickly.concertservice.exception.detail;
+
+public class ConcertSearchException extends RuntimeException {
+    public ConcertSearchException(String message) {
+        super(message);
+    }
+}
