+++ concert-service/src/main/java/com/eatpizzaquickly/concertservice/exception/detail/ElasticsearchIndexException.java
@@ -1,0 +1,8 @@
+package com.eatpizzaquickly.concertservice.exception.detail;
+
+public class ElasticsearchIndexException extends RuntimeException {
+
+    public ElasticsearchIndexException(String message) {
+        super(message);
+    }
+}
